added changelog and refactored code to separate cli
